Rename file too
