+++ pumpkin-protocol/src/server/play/interact.rs
@@ -1,7 +1,7 @@
 use bytes::Buf;
 use pumpkin_data::packet::serverbound::PLAY_INTERACT;
 use pumpkin_macros::server_packet;
-use pumpkin_util::math::vector3::Vec3;
+use pumpkin_util::math::vec3::Vec3;
 
 use crate::{
     ServerPacket, VarInt,
