+++ pumpkin-protocol/src/server/play/player_position.rs
@@ -1,6 +1,6 @@
 use pumpkin_data::packet::serverbound::PLAY_MOVE_PLAYER_POS;
 use pumpkin_macros::server_packet;
-use pumpkin_util::math::vector3::Vec3;
+use pumpkin_util::math::vec3::Vec3;
 
 #[derive(serde::Deserialize)]
 #[server_packet(PLAY_MOVE_PLAYER_POS)]
