+++ pumpkin-world/src/chunk/anvil.rs
@@ -146,7 +146,7 @@ impl ChunkReader for AnvilChunkFormat {
     fn read_chunk(
         &self,
         save_file: &LevelFolder,
-        at: &pumpkin_util::math::vector2::Vec2<i32>,
+        at: &pumpkin_util::math::vec2::Vec2<i32>,
     ) -> Result<super::ChunkData, ChunkReadingError> {
         let region = (at.x >> 5, at.z >> 5);
 
@@ -230,7 +230,7 @@ impl ChunkWriter for AnvilChunkFormat {
         &self,
         chunk_data: &ChunkData,
         level_folder: &LevelFolder,
-        at: &pumpkin_util::math::vector2::Vec2<i32>,
+        at: &pumpkin_util::math::vec2::Vec2<i32>,
     ) -> Result<(), super::ChunkWritingError> {
         let region = (at.x >> 5, at.z >> 5);
 
@@ -499,7 +499,7 @@ impl AnvilChunkFormat {
 
 #[cfg(test)]
 mod tests {
-    use pumpkin_util::math::vector2::Vec2;
+    use pumpkin_util::math::vec2::Vec2;
     use std::fs;
     use std::path::PathBuf;
     use temp_dir::TempDir;
