+++ pumpkin-world/src/generation/generic_generator.rs
@@ -1,5 +1,5 @@
 use noise::{NoiseFn, Perlin};
-use pumpkin_util::math::vector2::Vec2;
+use pumpkin_util::math::vec2::Vec2;
 
 use crate::{
     WORLD_LOWEST_Y,
