+++ pumpkin/src/entity/mob/mod.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use async_trait::async_trait;
 use pumpkin_data::entity::EntityType;
-use pumpkin_util::math::vector3::Vec3;
+use pumpkin_util::math::vec3::Vec3;
 use tokio::sync::Mutex;
 use zombie::Zombie;
