+++ pumpkin/src/entity/player.rs
@@ -55,8 +55,8 @@ use pumpkin_util::{
         boundingbox::{BoundingBox, EntityDimensions},
         experience,
         position::BlockPos,
-        vector2::Vec2,
-        vector3::Vec3,
+        vec2::Vec2,
+        vec3::Vec3,
     },
     permission::PermissionLvl,
     text::TextComponent,
