+++ pumpkin/src/server/mod.rs
@@ -10,8 +10,8 @@ use pumpkin_protocol::{ClientPacket, client::config::CPluginMessage};
 use pumpkin_registry::{DimensionType, Registry};
 use pumpkin_util::math::boundingbox::{BoundingBox, EntityDimensions};
 use pumpkin_util::math::position::BlockPos;
-use pumpkin_util::math::vector2::Vec2;
-use pumpkin_util::math::vector3::Vec3;
+use pumpkin_util::math::vec2::Vec2;
+use pumpkin_util::math::vec3::Vec3;
 use pumpkin_util::text::TextComponent;
 use pumpkin_world::block::registry::Block;
 use pumpkin_world::dimension::Dimension;
